pyproject.toml
[project]
name = "projeto1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]
